backend/pyproject.toml
[tool.poetry]
name = "fastapiproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.21"
celery = "^5.3.4"
passlib = "^1.7.4"
fastapi-sqlalchemy = "^0.2.1"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
